requirements.txt
bottle
gunicorn
psycopg2-binary
